standardize use of @api_view, misc CSRF/auth fixes
