+++ dww_backend/api/views/patient_reminder_views.py
@@ -2,7 +2,6 @@
 from django.http import JsonResponse
 from django.contrib.auth import authenticate, logout, login as django_login
 from django.views.decorators.csrf import csrf_exempt
-from django.http import JsonResponse
 from api.models import *
 from api.forms import *
 from api.serializers import *
@@ -17,24 +16,7 @@ Note: All patient-facing APIs should use rest_framework's JWT authentication
 """
 
 
-@csrf_exempt 
-@api_view(['POST'])
-@authentication_classes([JWTAuthentication])
-@permission_classes([IsAuthenticated])
-def add_reminder(request):
-    try:
-        user = request.user
 
-        serializer = ReminderSerializer(data=request.data)
-        if not serializer.is_valid():
-            return Response({'error': serializer.errors}, status=400)
-        
-        PatientReminder.objects.create(patient=user, time=serializer.validated_data['time'], days=serializer.validated_data['days'])
-        return JsonResponse({'message': 'Reminder added successfully'}, status=201)
-    except Exception as e:
-        return JsonResponse({'error': str(e)}, status=500)
-
-@csrf_exempt 
 @api_view(['GET'])
 @authentication_classes([JWTAuthentication])
 @permission_classes([IsAuthenticated])
@@ -46,7 +28,24 @@ def get_reminders(request):
     except Exception as e:
         return JsonResponse({'error': str(e)}, status=500)
 
-@csrf_exempt   
+
+@api_view(['POST'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
+def add_reminder(request):
+    try:
+        user = request.user
+        serializer = ReminderSerializer(data=request.data)
+        if not serializer.is_valid():
+            return Response({'error': serializer.errors}, status=400)
+        
+        PatientReminder.objects.create(patient=user, time=serializer.validated_data['time'], days=serializer.validated_data['days'])
+        return JsonResponse({'message': 'Reminder added successfully'}, status=201)
+    
+    except Exception as e:
+        return JsonResponse({'error': str(e)}, status=500)
+
+
 @api_view(['PUT'])
 @authentication_classes([JWTAuthentication])
 @permission_classes([IsAuthenticated])
@@ -54,7 +53,6 @@ def save_reminder(request):
     try:
         user = request.user
         reminder_id = request.data.get('id')
-
         serializer = ReminderSerializer(data=request.data)
         if not serializer.is_valid():
             return Response({'error': serializer.errors}, status=400)
@@ -66,28 +64,26 @@ def save_reminder(request):
         
         reminder.time = serializer.validated_data['time']
         reminder.days = serializer.validated_data['days']
-
         reminder.save()
-
         return JsonResponse({'message': 'Reminder saved successfully'}, status=201)
+    
     except Exception as e:
         return JsonResponse({'error': str(e)}, status=500)
 
-@csrf_exempt   
+
 @api_view(['DELETE'])
 @authentication_classes([JWTAuthentication])
 @permission_classes([IsAuthenticated])
 def delete_reminder(request, id):
     try:
         user = request.user
-        
         try:
             reminder = PatientReminder.objects.get(id=id, patient=user)
         except PatientReminder.DoesNotExist:
             return JsonResponse({'error': 'Reminder not found'}, status=404)
         
         reminder.delete()
-
         return JsonResponse({'message': 'Reminder deleted successfully'}, status=201)
+    
     except Exception as e:
         return JsonResponse({'error': str(e)}, status=500)
